default attr_timeout=0
We don't want FUSE to cache attributes because they might change as we
gather information about files.

Without this option, a "cat" on a log file, for instance, doesn't work
the first time because FUSE does a LOOKUP first and remembers the file
to have size 0, and consequently refuses to let programs read from it.